premier depot

+++ src/main/java/com/bocobi/back/repos/AbonnementRepos.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.bocobi.back.models.Abonnement;
 
 
-public interface AbonnementRepos extends JpaRepository<Abonnement, Integer> {
+public interface AbonnementRepos extends JpaRepository<Abonnement, Long> {
 
 }
