+++ src/main/java/com/bocobi/back/repos/PatientRepos.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bocobi.back.models.Patient;
 
-
-public interface PatientRepos extends JpaRepository<Patient, Integer> {
+public interface PatientRepos extends JpaRepository<Patient, Long> {
 
 }
